Systems Project/requirements.txt
dash
dash-bootstrap-components
easyfsl
fastapi
matplotlib
numpy
opencv-python-headless
pandas
Pillow
torch
torchvision
uvicorn